Android article design reviews (#777)
* fix: Check if byline present before adding byline row

* Update padding on standfirst

* Updated top padding for date publication

* Lineheight and padding for byline and datepub

* Update android article snapshots

* Update flag snapshots for android

* Move styles to a style folder in article-flag
