+++ packages/article-flag/style/index.android.js
@@ -1,0 +1,11 @@
+import sharedStyles from "./shared";
+
+const styles = {
+  ...sharedStyles,
+  title: {
+    ...sharedStyles.title,
+    fontSize: 14
+  }
+};
+
+export default styles;
