+++ packages/article-flag/style/index.js
@@ -1,0 +1,3 @@
+import sharedStyles from "./shared";
+
+export default sharedStyles;
