+++ packages/article-flag/style/shared.js
@@ -1,0 +1,20 @@
+import { fonts } from "@times-components/styleguide";
+
+const styles = {
+  view: {
+    flexDirection: "row",
+    alignItems: "center"
+  },
+  diamond: {
+    marginRight: 5,
+    marginBottom: 1
+  },
+  title: {
+    fontFamily: fonts.bodyRegularSmallCaps,
+    fontSize: 12,
+    fontWeight: "400",
+    letterSpacing: 1.4
+  }
+};
+
+export default styles;
