+++ packages/article/styles/article-header/index.android.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { fontSizes, spacing } from "@times-components/styleguide";
+import { fontSizes } from "@times-components/styleguide";
 import globalStyle from "../shared";
 import sharedStyles from "./shared";
 
@@ -14,7 +14,7 @@ const androidStyles = {
   standFirst: {
     ...sharedStyles.standFirst,
     lineHeight: 29,
-    paddingBottom: spacing(3),
+    paddingBottom: 12,
     marginTop: -7
   }
 };
